Move configuration out of Blueprint
This centralizes configuration so all of the environment access is done
in the same place in run.py and the Blueprint can use flask.current_app
to access the config values. Moving the environment access out of the
Blueprint makes it easier to swap the source of the config e.g. to a
file or to something like Vault, and keeps the Blueprint reusable.